Test retry handler with multiple workers
